requirements.txt
# Core dependencies
google-api-python-client
google-auth-httplib2
google-auth-oauthlib




# Scheduling & time handling
APScheduler

# Database (PostgreSQL + ORM)
sqlalchemy
psycopg2-binary
psycopg2

# Environment & config
python-dotenv

# Web framework
fastapi
uvicorn

python-multipart
